chore: prepare knx-pico v0.2.1 for publication
- Fix trivial numeric cast warnings in macros
- Fix unused mut warning in tests
- Add missing documentation for pico_log macro
- Move src/main.rs to examples/knx_main_application.rs for documentation
- Update build.rs to only include example link arguments
- Update Cargo.toml exclude list to include necessary files for examples
- Include build.rs, memory.x, and configuration.rs in published package

+++ src/macros.rs
@@ -80,7 +80,7 @@ macro_rules! ga {
         };
 
         // Calculate raw address: MMMMMMMM MMMMSSSSSSSSS (5 bits main, 3 bits middle, 8 bits sub)
-        const RAW: u16 = (($main as u16 & 0x1F) << 11) | (($middle as u16 & 0x07) << 8) | ($sub as u16 & 0xFF);
+        const RAW: u16 = (($main & 0x1F) << 11) | (($middle & 0x07) << 8) | ($sub & 0xFF);
         $crate::addressing::GroupAddress::from(RAW)
     }};
 }
